refactor: rename database-related modules and update import paths
