+++ tests/database/connection/test_connector.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import MagicMock, patch
 
-from src.shared.database.connector import DatabaseConnector
+from src.database.connection.connector import DatabaseConnector
 
 
 class TestDatabaseConnector(unittest.TestCase):
@@ -25,17 +25,17 @@ class TestDatabaseConnector(unittest.TestCase):
             "port": 22,
         }
 
-    @patch("src.shared.database.connector.PRD", "EnduranceProject")
+    @patch("src.database.connection.connector.PRD", "EnduranceProject")
     def test_is_prd_environment(self):
         assert self.database_connector._is_prd_environment() == True
 
-    @patch("src.shared.database.connector.PRD", "")
+    @patch("src.database.connection.connector.PRD", "")
     def test_is_local_environment(self):
         assert self.database_connector._is_prd_environment() == False
 
-    @patch("src.shared.database.connector.PRD", "EnduranceProject")
-    @patch("src.shared.database.connector.create_engine")
-    @patch("src.shared.database.connector.scoped_session")
+    @patch("src.database.connection.connector.PRD", "EnduranceProject")
+    @patch("src.database.connection.connector.create_engine")
+    @patch("src.database.connection.connector.scoped_session")
     def test_can_create_remote_engine_with_sucess(self, mock_scoped_session, mock_create_engine):
         mock_mysql_credentials = MagicMock()
         mock_ssh_credentials = MagicMock()
@@ -44,8 +44,8 @@ class TestDatabaseConnector(unittest.TestCase):
         mock_scoped_session.assert_called_once()
         mock_create_engine.assert_called_once()
 
-    @patch("src.shared.database.connector.PRD", "")
-    @patch("src.shared.database.connector.DatabaseConnector._get_local_engine_url")
+    @patch("src.database.connection.connector.PRD", "")
+    @patch("src.database.connection.connector.DatabaseConnector._get_local_engine_url")
     def test_exception_when_create_local_engine(self, mock_get_local_engine_url):
         mock_mysql_credentials = MagicMock()
         mock_ssh_credentials = MagicMock()
@@ -56,7 +56,7 @@ class TestDatabaseConnector(unittest.TestCase):
 
         self.assertEqual(str(context.exception), "Test exception")
 
-    @patch("src.shared.database.connector.sshtunnel.SSHTunnelForwarder")
+    @patch("src.database.connection.connector.sshtunnel.SSHTunnelForwarder")
     def test_can_get_local_engine_url(self, mock_SSHTunnelForwarder):
         mock_tunnel = MagicMock()
         mock_tunnel.local_bind_port = 12345
@@ -80,7 +80,7 @@ class TestDatabaseConnector(unittest.TestCase):
             self.database_connector._construct_engine_url(self.mock_mysql_keys, "localhost:12345"),
         )
 
-    @patch("src.shared.database.connector.sshtunnel.SSHTunnelForwarder")
+    @patch("src.database.connection.connector.sshtunnel.SSHTunnelForwarder")
     def test_exception_when_get_local_engine_url(self, mock_SSHTunnelForwarder):
         mock_mysql_keys = MagicMock()
         mock_ssh_keys = MagicMock()
@@ -93,7 +93,7 @@ class TestDatabaseConnector(unittest.TestCase):
         self.assertEqual(str(context.exception), "Test exception")
         self.database_connector.logger.error.assert_called_once_with("Failed to start SSH tunnel: Test exception")
 
-    @patch("src.shared.database.connector.DatabaseConnector._create_engine")
+    @patch("src.database.connection.connector.DatabaseConnector._create_engine")
     def test_can_get_new_session(self, mock_create_engine):
         mock_mysql_keys = MagicMock()
         mock_ssh_keys = MagicMock()
@@ -103,7 +103,7 @@ class TestDatabaseConnector(unittest.TestCase):
         mock_create_engine.assert_called_once_with(mock_mysql_keys, mock_ssh_keys)
         self.database_connector.logger.info.assert_called_once_with("Session is not initialized. Creating engine.")
 
-    @patch("src.shared.database.connector.DatabaseConnector._create_engine")
+    @patch("src.database.connection.connector.DatabaseConnector._create_engine")
     def test_can_get_existing_session(self, mock_create_engine):
         self.database_connector.Session = MagicMock()
